Implement RecipeDetailView
Current implementation is very basic, just a label for title
and a button which opens a URL in the user's default
browser. There's a bunch of weird inconsistencies left
throughout the code, but it's enough to move forward for
now.

Shoutouts to StackOverflow users for the Linux workaround
in RecipeDetailView.handleLink.