Extract browser-launching code into standalone class
I'm not 100% on the static methods in here, but
Desktop.getDesktop().browse() is a static method call
(or at least getDesktop() is.) And I don't expect
polymorphism to be a problem.

That said, we could move to a polymorphic form; maybe a
WindowsBrowserLauncher and a LinuxBrowserLauncher? feels
a little overkill though.