Rename getApiData to loadApiKeys
Feels better, but I'm not sure it's perfect.

Side note, I just realized the the last few
commits have been made under the wrong name and
email. I (think I've) changed the most recent
few, but it doesn't appear to be possible to
fix the ones that were already pushed.