Add MenuStateChangeIntermediary
This commit represents a major change in how the
application state is handled. Previously, each major view
class needed to manually show the other in order to switch,
which required that the Schedule be passed between them.
Now, showing new views is handled by the
MenuStateChangeIntermediary, which also holds the canonical
reference to the Schedule.

There's a variety of weird deep-copy things going on to
make this make sense, and there's still cases (namely
ScheduleEditorBox) where the Schedule is accessed by
reference. The search dialog is also not currently managed
by the intermediary, though it probably could be.