Add support for nutrient data to WebRecipe
This includes a great deal of changes to the serialization
tests and a bunch of new files and stuff.

I'm not happy with the initializeMenu() methods in those
tests, as it happens, but I'm not going to attempt to
refactor them right now.

I'm also ignoring a couple of warnings in WebRecipe about
unused fields; we'll need them to display the data in the
UI.