Suppress IDE warning
IDEA was warning about an unused schema; I don't fully
understand how SVG works, but I don't want to modify this
file. So, I'm suppressing the warning.

I want to avoid modifying it for legal reasons, to be
clear. It may not actually matter, but I'd rather be safe
than sued.