Refactor Schedule
Most notably, converted it to a record. This is so that
we can use the autogenerated equals method mostly. I'm
slightly concerned that this might break encapsulation
on menus a bit, but it might also be fine. I've had enough
problems dealing with equals methods that this seems like
a reasonable move, at least.